Start using .env files
Instead of commiting a serverles..yml.dist file, I realized it makes more sense to have a .env file. This way, I'm commiting here a .env.dist where the secret variables will be saved.

The serverles.yml still does not read correctly the .env file when deploying, but this can be fixed later.