Implements user endpoint
The first real endpoint was implemented on this commit to fetch users, based on the uuid. This gave me a lot of ideas on how to use express and specially on the power of middlewares that will be used a lot in the future.

Now that it is possible to fetch for users, I will implement next the creation of measures and its model.